simplify code of branch documented-custom-validators
