--- src/main/java/sample/Person.java
@@ -1,104 +1,0 @@
-package sample;
-
-import javax.validation.constraints.Email;
-import javax.validation.constraints.Max;
-import javax.validation.constraints.Min;
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Pattern;
-import javax.validation.constraints.Size;
-import javax.xml.bind.annotation.XmlAccessType;
-import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlRootElement;
-
-import org.hibernate.validator.constraints.CreditCardNumber;
-
-import sample.customvalidations.DateTimeType;
-import sample.customvalidations.LocalDateTimeFormat;
-
-@XmlRootElement(name = "person")
-@XmlAccessorType(XmlAccessType.FIELD)
-public class Person {
-    private long id;
-    private String firstName;
-    @NotNull
-    @NotBlank
-    private String lastName;
-    @Pattern(regexp = ".+@.+\\..+", message = "Please provide a valid email address")
-    private String email;
-    @Email()
-    private String email1;
-    @Min(18)
-    @Max(30)
-    private int age;
-    @CreditCardNumber
-    private String creditCardNumber;
-    
-    @LocalDateTimeFormat(pattern = "yyyyMMdd",   dateTimeType=DateTimeType.Date, message = "Invalid dateTimeField Format. It Should be in yyyyMMdd format")
-	private String registrationDate;
-	
-	public String getCreditCardNumber() {
-        return creditCardNumber;
-    }
-
-    public void setCreditCardNumber(String creditCardNumber) {
-        this.creditCardNumber = creditCardNumber;
-    }
-
-    public long getId() {
-        return id;
-    }
-
-    public void setId(long id) {
-        this.id = id;
-    }
-    public String getEmail1() {
-        return email1;
-    }
-
-    public void setEmail1(String email1) {
-        this.email1 = email1;
-    }
-
-    @Size(min = 2)
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
-
-    public String getLastName() {
-
-        return lastName;
-    }
-
-    public void setLastName(String lastName) {
-        this.lastName = lastName;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public int getAge() {
-        return age;
-    }
-
-    public void setAge(int age) {
-        this.age = age;
-    }
-    
-    public String getRegistrationDate() {
-		return registrationDate;
-	}
-	
-	public void setRegistrationDate(String registrationDate) {
-		this.registrationDate = registrationDate;
-	}
-}
