--- src/main/java/sample/PersonController.java
@@ -1,16 +1,0 @@
-package sample;
-import javax.validation.Valid;
-
-import org.springframework.http.MediaType;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class PersonController {
-    @RequestMapping(path = "/person", method = RequestMethod.POST , consumes = {MediaType.APPLICATION_JSON_VALUE, MediaType.APPLICATION_XML_VALUE})
-    public Person person(@Valid @RequestBody Person person) {
-        return person;
-    }
-}
