+++ src/main/java/sample/config/DateTimeFormatData.java
@@ -1,0 +1,14 @@
+package sample.config;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+class DateTimeFormatData {
+
+	private String pattern;
+
+	private String dateTimeType;
+
+}
