+++ src/main/java/sample/customvalidations/DateTimeType.java
@@ -1,6 +1,7 @@
 package sample.customvalidations;
 
 public enum DateTimeType {
+
 	DateTime,
 	Date,
 	Time
