+++ src/main/java/sample/customvalidations/LocalDateTimeFormat.java
@@ -1,7 +1,5 @@
 package sample.customvalidations;
 
-import static java.lang.annotation.RetentionPolicy.RUNTIME;
-
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -10,21 +8,23 @@ import java.lang.annotation.Target;
 import javax.validation.Constraint;
 import javax.validation.Payload;
 
-@Target({ElementType.FIELD, ElementType.METHOD, ElementType.PARAMETER, 
-	ElementType.ANNOTATION_TYPE })
+import static java.lang.annotation.RetentionPolicy.RUNTIME;
+
+@Target({ ElementType.FIELD, ElementType.METHOD, ElementType.PARAMETER,
+		ElementType.ANNOTATION_TYPE })
 @Retention(RUNTIME)
 @Constraint(validatedBy = LocalDateTimeValidator.class)
 @Documented
 public @interface LocalDateTimeFormat {
 
-    String message() default "{message.key}";
+	String message() default "{message.key}";
 
-    Class<?>[] groups() default { };
+	Class<?>[] groups() default {};
 
-    Class<? extends Payload>[] payload() default { };
+	Class<? extends Payload>[] payload() default {};
 
-    String pattern();
-    
-    DateTimeType dateTimeType() default DateTimeType.DateTime;
+	String pattern();
+
+	DateTimeType dateTimeType() default DateTimeType.DateTime;
 
 }
