+++ src/main/resources/application.properties
@@ -4,4 +4,4 @@ springdoc.swagger-ui.show-extensions=true
 springdoc.swagger-ui.show-common-extensions=true
 server.error.include-message=always
 server.error.include-binding-errors=always
-
+springdoc.use-fqn=true
